requirements.txt
oracledb==2.4.1
PyYAML==6.0.2
pandas==2.2.3